requirements.txt
transformers
langchain
openai
streamlit
python-dotenv
requests
